Feature: Consistency checks

+++ pyhope/mesh/topology/mesh_topology.py
@@ -615,7 +615,7 @@ def prism_faces(order: int) -> tuple[np.ndarray, ...]:
 # Dummy function for hexahedral elements
 @cache
 def split_hex_to_hex(order: int) -> list[tuple[int, ...]]:
-    nodes = np.arange((order + 1) ** 3)
+    nodes = np.arange((order + 1) ** 3, dtype=int)
     return [tuple(nodes)]
